+++ src/common/cout970/UltraTech/machines/containers/CVDcontainer.java
@@ -45,7 +45,7 @@ public class CVDcontainer extends Container{
     {
         super.addCraftingToCrafters(par1ICrafting);
         par1ICrafting.sendProgressBarUpdate(this, 0, tileEntity.progres);
-        par1ICrafting.sendProgressBarUpdate(this, 1, tileEntity.getEnergy());
+        par1ICrafting.sendProgressBarUpdate(this, 1, (int)tileEntity.getEnergy());
         par1ICrafting.sendProgressBarUpdate(this, 2, tileEntity.speed);
     }
